==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month:int) -> Path:
    """"""
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200 :
        path = f'../data/rides_{year}-{month:02d}.parquet'
        open(path, "wb").write(response.content)
        return path
    else :
        raise Exception(f'{URL} is not available')


download_one_file_of_raw_data(year= 2022, month=1)

import pandas as pd

rides = pd.read_parquet('../data/rides_2022-01.parquet')

rides.head(20)

# we keep the columns we want

rides = rides[['tpep_pickup_datetime', 'PULocationID']]


rides.head()

rides.rename(columns={'tpep_pickup_datetime' : 'pickup_datetime', 'PULocationID' : 'pickup_location_id'}, inplace=True)

rides.head()

# let's check the columns

rides['pickup_datetime'].describe()

# we should remove some rows because we just want the data in january

type(rides['pickup_datetime'][0])

# the date's type is Timestamp and not datetime !

rides = rides[ (rides['pickup_datetime'] <= '2022-02-01') & (rides['pickup_datetime'] >= '2022-01-01') ]

rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
