==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
# We want to transform the raw data filtered (01/2022) into timeseries data

import pandas as pd

# we add the 'r' so that python can read the parquet file

rides = pd.read_parquet(r'C:\Users\hiche\taxi_demand_predictor\data\transformed\validated_rides_2022_01.parquet')

rides.head()

# We create a new feature based on the datetime, rounded to the closest hour

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')

rides['pickup_hour']

# We aggregate the data for each hour of each day

agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0:'rides'}, inplace=True)
agg_rides

# we want a "0" in the rows where there isn't rides for the pickup_location_id X (where there is no hours assigned)
# that is to say : we add 0 to the missing rows
# we do the operations for each location_id
# tqdm is a package that adds a progress bar to looping operations
# there's a way to do it : 

idx = pd.date_range('09-01-2013', '09-30-2013')
#print(idx)

s = pd.Series({'09-02-2013': 2,
               '09-03-2013': 10,
               '09-06-2013': 5,
               '09-07-2013': 1})
s.index = pd.DatetimeIndex(s.index)
#print(s)

s = s.reindex(idx, fill_value=0)
#print(s)

# we want a "0" in the rows where there isn't rides for the pickup_location_id X
# that is to say : we add 0 to the missing rows
# we do the operations for each location_id
# tqdm is a package that adds a progress bar to looping operations

from tqdm import tqdm


def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame :
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    output = pd.DataFrame()


    for location_id in tqdm(location_ids) :

        # keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        # quick way to add missing dates with 0 in a Series
        # taken from https://stackoverflow.com/a/19324591
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)


        # add back 'location_id' columns
        agg_rides_i['pickup_location_id'] = location_id
        

        output = pd.concat([output, agg_rides_i])

    # move the purchase_day from the index to a dataframe column
    output = output.reset_index().rename(columns={'index' : 'pickup_hour'})

    return output

agg_rides_all_slot = add_missing_slots(agg_rides=agg_rides)


# plot the data

# to make my code readable
from typing import Optional, List

# plotly for interactivity (zoom in and out, details)
import plotly.express as px

def plot_rides(rides : pd.DataFrame, locations : Optional[List[int]] = None) :
    """
    Plot time-series data
    """

    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(rides_to_plot, x = 'pickup_hour', y = 'rides', color = 'pickup_location_id', template='none')
    fig.show()


plot_rides(agg_rides_all_slot, locations=[43])

agg_rides_all_slot.to_parquet('../data/transformed/ts_data_2022_01.parquet')
