==> notebooks/03_transform_ts_data_into_features_and_target.ipynb <==
# We want to transform the timeseries data into features and target for ML

import pandas as pd

ts_data = pd.read_parquet('../data/transformed/ts_data_2022_01.parquet')

ts_data.head()

ts_data_one_location = ts_data.loc[ts_data.pickup_location_id == 43, :].reset_index(drop=True)

ts_data_one_location.info()
ts_data_one_location.head()

# 43 is Central Park
# 24 hours * 31 days = 744 rows

# we'll have x_1 to x_24, where x_1 = from index 0 to 24, target = 25, we want to predict the number of rides the next hour
# there will be 744 - n_features ROWS


# slice and slide

# slice

def get_cutoff_indices(
        data: pd.DataFrame,
        n_features: int,
        step_size: int
        ) -> list :
    
    stop_position = len(data) - 1

    # start the fist sub-sequence at index position 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []

    while subseq_last_idx <= stop_position :

        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))

        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size
    
    return indices


# we test the function, n_features = 24 for the hours
# we'll have x_1 to x_24, where x_1 = from index 0 to 24, target = 25, we want to predict the number of rides the next hour

n_features = 24
step_size = 1

indices = get_cutoff_indices(ts_data_one_location, n_features, step_size)

indices

# there is 744 - 25 (n_features) = 719 rows

# slide 

import numpy as np

n_examples = len(indices)
# = 719

x = np.ndarray(shape=[n_examples,n_features], dtype=np.float32)
y = np.ndarray(shape=[n_examples], dtype=np.float32)
pickup_hours = []

for i, idx in enumerate(indices) :
    x[i, :] = ts_data_one_location.iloc[idx[0] : idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1] : idx[2]]['rides'].values

    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])


# We see what X looks like

print(f'{x.shape=}')
print(f'{x=}')
print(f'{pickup_hours[:5]=}')

# we want X to be a dataframe

features_one_location = pd.DataFrame(x, columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))] )

features_one_location.shape

features_one_location.head(5)

# we want y to be a dataframe

targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])

targets_one_location.head()

# for each location for all locations

from tqdm import tqdm 

def transform_ts_data_into_features_and_target(
        ts_data: pd.DataFrame,
        input_seq_len: int,
        step_size: int,
        ) -> pd.DataFrame:
    """
    Slices and transposes data from time-series format into a (features, target)
    format that we can use to train Supervised ML models
    """
    
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    location_ids = ts_data['pickup_location_id'].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()

    for location_id in tqdm(location_ids) :
        # keep only ts data for this 'location_id'
        ts_data_one_location = ts_data.loc[ts_data.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        # pre-compute cutoff indices to split dataframe rows
        indices = get_cutoff_indices(ts_data_one_location, input_seq_len, step_size)

        # slice and transpose data into numpy arrays for features and targets
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples,input_seq_len), dtype=np.float32)
        y = np.ndarray(shape=(n_examples), dtype=np.float32)
        pickup_hours = []

        for i, idx in enumerate(indices) :
            x[i, :] = ts_data_one_location.iloc[idx[0] : idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[1] : idx[2]]['rides'].values

            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

        # numpy --> pandas
        features_one_location = pd.DataFrame(x, columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(input_seq_len))] )
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        # numpy --> pandas
        targets_one_location = pd.DataFrame(y, columns=[f'targets_rides_next_hour'])

        # concatenate results
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets['targets_rides_next_hour']



# one week of history = 24*7*1

features, targets = transform_ts_data_into_features_and_target(ts_data, input_seq_len=24*7*1, step_size=24)

print(f'{features.shape=}')
print(f'{targets.shape=}')
